// src/pages/main-menu/order/user/menu/accBalance/models/interfaces.ts
export interface AccBalProps {
  acctId?: number;
}

// Generated by https://quicktype.io

export interface AccBalList {
  balId: number;
  acctId: number;
  acctResId: number;
  subsId: number;
  custId: null;
  grossBal: number;
  reserveBal: number;
  consumeBal: number;
  ratingBal: number;
  billingBal: number;
  realBal: number;
  lastBal: number;
  usedBal: number;
  charge: number;
  seconds: number;
  preBalance: null;
  preSuttleBal: null;
  preEffDate: null;
  preExpDate: null;
  effDate: string;
  expDate: string;
  updateDate: string;
  minExpDate: null;
  ceilLimit: null;
  ceilLimitCharge: null;
  floorLimit: null;
  dailyCeilLimit: null;
  dailyFloorLimit: null;
  priority: null;
  balCode: null;
  routingId: null;
  absExpOffset: null;
  changeInitBal: null;
  initBal: number;
  lastRecharge: null;
  balUsed: number;
  operationType: null;
  acctResName: null;
  acctResComments: null;
  balFlags: null;
  isCurrency: string;
  isExpired: null;
  refAttr: null;
  checkMode: string;
  postCheckMode: string;
  balShareInWorkHourFlag: string;
  consumeBalCharge: null;
  balShareId: null;
  varCeilLimit: null;
  grossSubBal: null;
  effSeconds: null;
  acctResDto: null;
  balAcctItemTypeDtoList: null;
  balShareDtoList: null;
  balShareItemDtoList: null;
}